+++ Editor/ScriptableObjects/SelectionWindow/MultipleTypeSelectionWindow.cs
@@ -15,18 +15,20 @@
     /// </summary>
     internal class MultipleTypeSelectionWindow : TypeSelectionWindow
     {
-        private const float WindowWidth = 350f;
+        private const float WindowWidth = 300f;
 
         [SerializeField] private TypeReferenceWithBaseTypes[] _typeRefs;
 
         private Action<Type[]> _onTypesSelected;
         private SerializedObject _serializedObject;
         private ContentCache _contentCache;
+        private string[] _genericArgNames;
 
-        protected override void OnCreate(Action<Type[]> onTypesSelected, Type[][] genericParamConstraints)
+        protected override void OnCreate(Action<Type[]> onTypesSelected, string[] genericArgNames, Type[][] genericParamConstraints)
         {
             int typesCount = genericParamConstraints.Length;
             _onTypesSelected = onTypesSelected;
+            _genericArgNames = genericArgNames;
             _typeRefs = new TypeReferenceWithBaseTypes[typesCount];
 
             for (int i = 0; i < typesCount; i++)
@@ -39,6 +41,7 @@
                 };
             }
 
+            titleContent = new GUIContent("Choose Arguments");
             _serializedObject = new SerializedObject(this);
             _contentCache = new ContentCache();
             this.Resize(WindowWidth, GetWindowHeight(typesCount));
@@ -48,11 +51,12 @@
         protected override void OnGUI()
         {
             SerializedProperty typesArray = _serializedObject.FindProperty(nameof(_typeRefs));
+
             for (int i = 0; i < typesArray.arraySize; i++)
             {
                 EditorGUILayout.PropertyField(
                     typesArray.GetArrayElementAtIndex(i),
-                    _contentCache.GetItem($"Type Parameter #{i+1}"));
+                    _contentCache.GetItem(_genericArgNames[i]));
             }
 
             if ( ! GUILayout.Button("Create Asset"))
@@ -64,6 +68,7 @@
             }
             else
             {
+                Close();
                 _onTypesSelected(_typeRefs.CastToType());
             }
         }