+++ GenericUnityInternals~/GenericUnityEditorInternals/GenericHeaderEditor.cs
@@ -2,8 +2,10 @@
 {
     using System;
     using System.Collections.Generic;
+    using SolidUtilities.Helpers;
     using UnityEditor;
     using Util;
+    using Object = UnityEngine.Object;
 
     /// <summary>
     /// An extension of Editor that changes name of <see cref="GenericScriptableObject"/> assets in the Inspector header.
@@ -11,7 +13,7 @@
     /// </summary>
     public class GenericHeaderEditor : Editor
     {
-        private static readonly Dictionary<UnityEngine.Object, string> _targetTitlesCache = new Dictionary<UnityEngine.Object, string>();
+        private static readonly Dictionary<TargetInfo, string> _targetTitlesCache = new Dictionary<TargetInfo, string>();
         private static readonly Dictionary<Type, string> _typeNamesCache = new Dictionary<Type, string>();
 
         internal override string targetTitle => GetTitle();
@@ -30,7 +32,9 @@
 
         private string GetOneTitle(Type genericType)
         {
-            if (_targetTitlesCache.TryGetValue(target, out string title))
+            var targetInfo = new TargetInfo(target);
+
+            if (_targetTitlesCache.TryGetValue(targetInfo, out string title))
                 return title;
 
             string typeName = GetTypeName(genericType);
@@ -40,7 +44,7 @@
                 return $"({typeName})";
 
             title = $"{ObjectNames.NicifyVariableName(target.name)} ({typeName})";
-            _targetTitlesCache.Add(target, title);
+            _targetTitlesCache.Add(targetInfo, title);
             return title;
         }
 
@@ -58,5 +62,40 @@
             _typeNamesCache.Add(genericType, typeName);
             return typeName;
         }
+    }
+
+    internal readonly struct TargetInfo : IEquatable<TargetInfo>
+    {
+        public readonly Object Target;
+        public readonly string Name;
+
+        public TargetInfo(Object target)
+        {
+            Target = target;
+            Name = target.name;
+        }
+
+        public override bool Equals(object obj)
+        {
+            return obj is TargetInfo info && Equals(info);
+        }
+
+        public bool Equals(TargetInfo p)
+        {
+            return (Target == p.Target) && (Name == p.Name);
+        }
+
+        public override int GetHashCode()
+        {
+            int hash = 17;
+            // ReSharper disable once Unity.NoNullPropagation
+            hash = hash * 23 + Target?.GetHashCode() ?? 0;
+            hash = hash * 23 + Name.GetHashCode();
+            return hash;
+        }
+
+        public static bool operator ==(TargetInfo lhs, TargetInfo rhs) => lhs.Equals(rhs);
+
+        public static bool operator !=(TargetInfo lhs, TargetInfo rhs) => ! lhs.Equals(rhs);
     }
 }